Fix cinderlib Python package
Setuptools gives 2 errors with the current cinderlib Python package:

- Project's description-file has an unsupported directive.
- test-requirements has a github reference.

We fix the first issue by removing the git repository from the test
requirements and adding it to the Tox file while adding cinder to the
required-projects to allow cross-repo denedencies.

We also add os-brick since we are currently also very dependent on it on
the nos_brick.py code.

For the second issue we create a new DESCRIPTION.rst file that doesn't
include the "raw" directive unsupported by PyPi.

This patch will allow us to add the publish-to-pypi job to the
project-config repository.

Required-By: I9d8d9f7d802a27c3bfe3dbc0ee84cc24eee24ebc
Change-Id: Id029cedef513753700a10bcae761fdc1807964e1