Fix Volume snapshots and connections on deletion
In some cases deleted snapshots and connections are not removed form the
cached lists, so we end up with a non existent element in the
"snapshots" or "connections" list that should not be there.

The reason why this happens is because we are looking for specific
instances instead of searching by the ID of the resource, and we could
have different instances in the list than the one we are looking for.

This patch ensure we are always looking for the elements to remove using
the ID.

Closes-Bug: #1819709
Change-Id: Ieb14e63c3de2f74f5d239a5174aeda9ee5c8fca1