Run functional tests with memory persistence
We are currently running functional tests with the memory DB persistence
plugin to ensure that the migrations run successfully, but this means
that we could be hiding issues on drivers that are not 100% compatible
with the persistence plugin mechanism, because the DB plugin still has
the SQLAlchemy implementation behind, which other plugins don't.

This patch first initializes cinderlib with the memory DB plugin to test
the DB migrations, and then replaces it with the memory plugin to ensure
driver compatibility by default.

We can still run the tests using the Memory DB persistence model setting
the CL_FTEST_MEMORY_PERSISTENCE environmental variable to false.

To ensure we are constantly testing both we run the LVM gate job with
Memory DB persistence and the Ceph gate with Memory persistence.

Related-Bug: #1821898
Change-Id: Id98c827ffadf2c47ff835afa8b0f83dddacb54fa