Fix get_connector_properties
We get exception oslo_privsep.daemon.FailedToDropPrivileges when calling
cinderlib.get_connector_properties.

This is because we make it point to OS-Brick's original method, which we
later replace when we do the init in nos-brick.  So cinderlib always
call the replaced method, while cinderlib users will be calling the
original OS-Brick method that hasn't replaced the privsep code.

This patch fixes this by assigning our get_connector_properties method
to the cinderlib method during the nos-brick initialization.

We also have a tiny doc change where we were saying connection instead
of connector, and some cosmetic changes in nos_brick to improve
readability, since we were using acronyms bgcp and bcp, which we now
replace for some more meaningful names.

Closes-Bug: #1822976
Change-Id: I627289418d52bdab826d1cf1a941e5671aaf2239