Add validation and validated documentation
This patch adds documentation related to the validation of Cinder
drivers for cinderlib usage.

With a guide on how to manually validate a driver using DevStack, and
how to add cinderlib functional tests to gate jobs.

Also presents a list of cinderlib validated drivers with their
configuration to serve as a reference for developers.

Depends-On: I25c19cf1f7da6881807e729a326cd94d07b4c42d
Depends-On: Ia532b3fcfbecee7c6dc7aa302e6d90a8067b7ff0
Change-Id: I2c184af9e48c3b2f4af7182fb000272f9e162be4