Fake unused packages
Many packages that are automatically imported when loading cinder
modules are only used for normal Cinder operation and are not necessary
for cinderlib's execution.

One example of this happening is when cinderlib loads a Cinder module to
get configuration options but won't execute any of the code present in
that module.

This patch fakes these unnecessary packages, providing faster load
times, reduced footprint, and the possibility for distributions to
create a cinderlib package or containers with up to 40% fewer
dependencies.

Change-Id: If577b9163d4e942b701db4b2a47cd66e6bd17b6f