zuul: change the cinderlib test runner into a role
A role can be easily reused elsewhere (for example by cinder
and cinder-tempest-plugins jobs).

The role is a wrapper around the default 'tox' role.
In order to make it working, CL_FTEST_CFG needs to be
overwritable completely, otherwise it is not possible
for tests to set a full path.

Finally, whitelist stestr to avoid tox warnings.

Change-Id: I274dd1da499e320e6d63af750cacb25e01df7561