Fix cloning metadata storage
When cloning volumes we are not properly storing the new volume's
metadata.

For one we don't store the source volume id because we use source_vol_id
instead of source_volid.

Then we also don't properly store the new state, so when we clone an
in-use volume our new volume will also be in-use.

This patch fixes the clone feature to work as expected.

Change-Id: If368f006b62c5133f6d671fcf4f4f8942d365ad8
Closes-Bug: #1849339
Closes-Bug: #1849828