Enhance extend functionality
Cinderlib's volume extend functionality would extend the volume in the
backend yet ignore any refresh of the view of the volume if it was
attached.

It also didn't provide a way to refresh the view of an extended volume
given a Connection instance.

This patch adds the funtionality to automatically refresh the view when
extending a locally attached volume as well as to do the refresh using a
Connection instance.

The refresh is done making the appropriate extend_volume call to
os-brick's connector.

The OS-Brick RBD connector doesn't have the method because there's
nothing for it to do, but to provide a consistent API we add the method
to our NOS-Brick to return the new size like other connectors do.

Change-Id: I04a29597e7515526fb8505c76c3d7e3026b2d9c8