Fix LVM extend volume
When we try to extend an attached LVM volume after retrieving the volume
metadata from the persistent storage we get a TargetUpdateFailed
exception cause by exception AttributeError: 'DB' object has no
attribute 'volume_attachment_get_all_by_volume_id'.

This is caused by the LVM driver checking volume_attachment which is not
populated, which triggers a lazy loading on the OVO which calls a method
in the DB layer that is not available in our fake DB instance.

This patch resolves this by replacing the get_all_by_volume_id method in
the VolumeAttachmentList OVO.

Close-Bug: #1852629
Change-Id: I7eade434b3d86e72678769f8ec929f66967d74e3