Relax checks on functional stats test
Functional test test_stats_with_creation has the limitation that it
requires exclusive use of the pool to run successfully.

This doesn't seem to be a safe assumption or even a reasonable request,
since there could be multiple cinderlib CI jobs running in parallel and
most CI systems will share the backend between Cinder CI and cinderlib
CI.

We still want to test that stats gathering does what it should do, or at
least that it looks that way, since we cannot test much given these
circumstances.

We modify the test to check that at least the values are not changed
when we don't refresh, and that they are changed when we refresh after
creating a volume.

Change-Id: I31fc6675ce3e079578f2c56f3e968ded8ab151c7