DBMS: Fix volume deletion with MySQL
When using the DBMS plugin with MySQL with default configuration (hard
DB row deletes) the deletion of a volume fails if the volume has
metadata or admin metadata.

The reason are the foreign key constraints from those 2 tables, because
we are not deleting their rows before deleting the volume row.

This doesn't fail with SQLite, but we leave rows behind.

With this patch we will delete all metadata and admin_metadata belonging
to the volume before we delete the volume row itself.

Closes-Bug: #1868148
Change-Id: I8712045752a7a150a3633451246ed992dee35972