NotLocal exception after refresh
If we locally attach a volume and we call the refresh method on it then
we cannot just call the detach method on the volume directly or we'll
get a `NotLocal` exception.

So this works:
   volume.attach()
   volume.detach()

But this doesn't:
   volume.attach()
   volume.refresh()
   volume.detach()

This is because we don't carry the `local_attach` attribute from our
current instance into the refresh.

Closes-Bug: #1868153
Change-Id: Iaaac5576d443b7917a46bb11458860a9651bd366