Fix CI jobs
Cinder added a new restriction to the snapshots in the DB in change
I47ff8115ae74e1a7ad41869159871ba614c388ac and now snapshot's
volume_type_id cannot be null.

This broke the cinderlib CI because one of our tests (test_set_snapshot
of the DBMS persistence plugin) didn't create the volume in the DB
(which can never happen in real use because it is saved on creation)
which is when the volume type gets assigned.

This patch saves the volume so that it gets a volume type and its ID can
be used by the Snapshot.

It also includes a workaround for the LVM job for
https://github.com/pypa/pip/issues/6264

Change-Id: I0993e2aab8511400247042c128dc3aa1a06486ac