Switch from mock to unittest.mock
We were relying on oslotest using the mock library and we were not
declaring it in our test-requirements.txt file, but now they are using
unittest.mock instead [1], so our unit tests are now failing because the
mock library is missing.

Switch to unittest.mock that is present in Python 3 to resolve the
problem.

[1]: https://review.opendev.org/#/c/715704/

Change-Id: If79500534aed2ab992c423f106088905edcb9d7f