RBD: Fix missing root_helper
When using the RBD driver and running the cinderlib process as a
non-root user within a container we'll get a "Command requested root,
but did not specify a root helper." error.

This error happens because the library is not passing the "root_helper"
parameter in a couple of "_execute" calls.

Trivial-Fix
Closes-Bug: #1885291
Change-Id: Idf664fdab3daf21b99215edfe3e2c053c3a9e85c