RBD: Ensure rbd_keyring_conf is available
With the security vulnerability (OSSN-0085) Cinder deprecated the
``rbd_keyring_conf`` option, but that's not a vulnerability for
cinderlib, since the user of the library already has access to the
keyring (otherwise the library couldn't read it).

Cinderlib requires this configuration option for the RBD driver to work
as expected, since cinderlib doesn't expect external user intervention
to make the keyring available on all the hosts.

In this patch we ensure that we report in the ``list_supported_drivers``
method option ``rbd_keyring_conf`` as an existing non-deprecated
configuration option, regardless if it's present (current Cinder code)
or if it has already been removed (which will happen soon).

Implements: blueprint rbd-ossn-0085-workaround
Change-Id: I7d3324e14494540f8c0d714682e1d0065c1735e3