Prevent duplicated backend names
With this patch we prevent us from having duplicated backend names for
different driver configurations during the same run.

If we pass the same name and configuration to Backend a second time we
get the Backend instance that was created on the first call.

If we pass the same name with a different configuration then we raise an
exception on the second call.

Change-Id: I87c14d236d7f2685315e4088db527090680c9523
Closes-Bug: #1886164