Fix and improve cross-project functional testing
Make sure stestr is executed using the python interpreter set inside
the virtualenv, otherwise the libraries installed there may not
found. This is due to the fact that (stestr default shebang points
to system python when installed through devstack and the functional
tox target uses that version thanks to sitepackages=True.
The sitepackages option  must be used to find a few system ceph/python
libraries not available on pypi.

Setting PATH is not really needed nowadays, so remove the custom
workaround. It is worth noting that the hack would not be needed anymore
and PATH could be set through setenv (extending {env:PATH}).

As cinderlib changes can break cinder and cinder-tempest-plugin,
and they indeed did it in the past, so ensure that everything works
by running cinder-tempest-plugin-lvm-lio-barbican.

Change-Id: I40ef42b4224521abed00b3b210418ec96c1116ea