Continue yoga development
As a cycle-with-intermediary release with trailing deliverable type,
cinderlib's master remains the yoga development branch until early in
the zed cycle when the cinderlib yoga release is made.  So update
tox.ini to test against cinder and os-brick stable/yoga instead of
their master branches, and restrict cinder and os-brick in
requirements.txt to yoga versions.

Change-Id: I8da1ddc38607eca378a503a961bf55a03566c5bc