Use stable/yoga override for CI jobs
cinderlib master is still Yoga development, but when zuul runs on a
master patch, zuul.branch is 'master', and this cascades throughout
the various jobs.  It's particularly evident for requirements, but
also affects cinder and os-brick (which are installed from source);
from what I can tell, an initial install is done to gather
dependencies, and this seems to be from master by default.  So
explicitly use override-checkout for requirements, cinder, and
os-brick on all jobs.

Why are we only noticing this now?  I think it hasn't been a problem
in the past because python 3.6 has been supported from Rocky through
Yoga, and is dropped from Zed, and hence running the cinderlib Yoga
development branch against master for other projects is encountering
some serious incompatibilities.  We probably should have been using
the stable branch in cinderlib's pre-release trailing phase all
along; it just hasn't mattered until now.

Change-Id: I82e2bbf7e0380e2bffbb44d394e9751e1578afcf