[ussuri-only] Remove cinderlib-lvm-functional job
The stable/ussuri branch of cinderlib has been in Extended Maintenance
phase since 2021-11-16.  The cinderlib-lvm-functional job has begun
failing, and while this has been fixed in more recent branches, those
branches contain other supporting changes that are not present in
this branch.  Given that Extended Maintenance phase allows a reduced
CI commitment [0], we propose dropping the job rather than investing
development effort in creating a new fix for a non-releaseable branch.

This patch also makes the cinderlib-ceph-functional job temporarily
non-voting as the fix for that job [1] is blocked by the cinderlib-
lvm-functional job.

[0] https://docs.openstack.org/project-team-guide/stable-branches.html#maintenance-phases
[1] https://review.opendev.org/c/openstack/cinderlib/+/850403/

Change-Id: If039438a8e8474644737e852d200de202a107979