Fix ValueError when running with zed cinder
When using the stable/zed cinder branch and zed upper-constraints,
we're getting a bunch of "ValueError: Backend named fake_backend
already exists with a different configuration" in unit tests.  The
backend name is set in the test setUp(), but it looks like isolation
isn't being respected, maybe because the Backend instantiation uses
a singleton pattern.  (Which doesn't, however, explain why this was
working previously.)  Address this by giving each test backend a
unique name.

Change-Id: I469b4b9dd9966be8477f3cefbd6284bdaf17ca17