Fix tox.ini
CI fails to run on Tox v4 complaining about spaces in pass_env and not
being able to run special_install.sh.

Those are easy to fix and this patch fixes them in tox.ini:

- Using multiple lines for the different environmental variable values
  in pass_env.

- Allowing running tools/special_install.sh and tools/generate_uc.sh as
  an external commands.

There is also an issue with the cinderlib entry points not being
available during unit tests.  This issues is not reproducible locally
even using the same tox and python version as the gate, so the patch
limits the tox version to v3 and prevents jobs from using v4 to unblock
the gate.

The patch also updates the values of the constraints branch from
antelope to 2023.1 since antelope doesn't exist and without this change
the patch cannot merge.

Change-Id: I14ea8821aeccc43b5ddcc83862d59b1bbf4094be