Don't use deprecated config opt name in unit test
With cinder change Iccfbe62406b6, the 'iscsi_secondary_ip_addresses'
opt was deprecated in favor of 'target_secondary_ip_addresses'.  This
broke a cinderlib unit test that checked to see if a mocked logger
was not called (which it now was, to log a deprecation message).  Use
the current option name in the test so that the test continues to
test what it was designed to test (i.e., cinderlib config loading,
not oslo.config option deprecation).

Change-Id: Ia243368a9d68aca3a717fb2be1da99bceb271e56