Use openstack-tox-py311
This addresses a TODO in .zuul.yaml.  The openstack-tox-py311 job
was added by change I725b94bee6601a.

The job runs on debian-bookworm, but we don't need to modify
cinderlib's bindep.txt because we don't include any packages not
recognized by debian.  Cinder does, however, so this patch depends
on the cinder patch that updates cinder's bindep.txt.

Depends-on: https://review.opendev.org/c/openstack/cinder/+/891645
Change-Id: Iff9d4079655c8c4b15ad54d6936a81d024aa2c74