Fix Zed branch CI
Zed is now call `unmaintained/zed` not `stable/zed`
This is requred for unmaintained/zed CI in Cinder.

NOTE(elod.illes): cinderlib is retired, hence the repository is empty
on master branch, causing issues with build-openstack-releasenotes job,
which should run on master. To unblock the gate this job (actually,
release-notes-jobs-python3 job template) needs to be dropped.

Change-Id: If769d9da92b724d40f47c9a57a75c021809c8850