CI: Fix unmaintained/2023.1 branch
Antelope is now called `unmaintained/2023.1`, not `stable/2023.1`. This
fix is requred for unmaintained/2023.1 CI in Cinder.

This is adapted from If769d9da92b724d40f47c9a57a75c021809c8850 which was
fixing the same issue for Zed.

NOTE(elod.illes): cinderlib is retired, hence the repository is empty on
master branch, causing issues with build-openstack-releasenotes job,
which should run on master. To unblock the gate this job (actually,
release-notes-jobs-python3 job template) needs to be dropped.

Change-Id: I04a9b621c12892ac1b0f53d97f43f38df160e7a0